#tic tac toe
This is a simple implementation of the classic Tic Tac Toe game in C. The game allows two players to take turns placing their markers (‘X’ or ‘O’) on a 3x3 grid. The objective is to get three of your markers in a row, either horizontally, vertically, or diagonally, before your opponent does.